clean up NestedAttributes, make it work for JsonAttribute::Model too
