+++ lib/json_attribute.rb
@@ -5,6 +5,7 @@ require "active_record/connection_adapters/postgresql_adapter"
 
 require 'json_attribute/record'
 require 'json_attribute/model'
+require 'json_attribute/nested_attributes'
 require 'json_attribute/record/query_scopes'
 
 # Dirty not supported on Rails 5.0
